requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
pyyaml==6.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
locust==2.17.0
python-dateutil==2.8.2
